refactor: ordertable 패키지 생성
