+++ src/main/java/kitchenpos/ui/response/TableResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.ui.response;
 
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.ordertable.OrderTable;
 
 import java.util.List;
 import java.util.stream.Collectors;
